+++ Models/StockItem.cs
@@ -1,7 +1,7 @@
 ﻿namespace StockTracking.Models;
 
 public abstract class StockItem : BaseEntity {
-    public int StockId { get; set; }
+    public long StockId { get; set; }
 
     public virtual Stock Stock { get; set; }
